requirements.txt
vk-api>=11.9.9
requests>=2.28.0
python-dotenv>=0.19.0
yadisk>=1.3.4
python-telegram-bot>=20.0
